techchain/forms/comment_create_form.py
from django import forms
from posts.models import Comment
from crispy_forms.helper import FormHelper

class CommentCreateForm(forms.ModelForm):
    class Meta:
        model = Comment
        fields = ['text']
        labels = {
            'text': '',
        }
        widgets = {
            'text': forms.Textarea(attrs={'class': 'form-control', 'rows': 2, 'placeholder': '💬 Comparte tu opinión...'}),
        }

        def __init__(self, *args, **kwargs):
            super().__init__(*args, **kwargs)
            self.fields['text'].label = ''
