requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.2.0
autobahn==24.4.2
Automat==24.8.1
bleach==6.2.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
crispy-bootstrap5==2024.10
cryptography==44.0.2
daphne==4.1.2
Django==5.1.3
django-appconf==1.2.0
django-cors-headers==4.9.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-js-asset==3.0.1
django-prose==2.1.0
django-select2==8.4.3
django-trumbowyg==1.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pillow==11.0.0
psycopg==3.2.12
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
redis==5.2.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uuid==1.30
webencodings==0.5.1
zope.interface==7.2
